+++ benchmark/streamly-benchmarks.cabal
@@ -136,7 +136,7 @@ library
 
 common bench-options
   import: compile-options, optimization-options, bench-depends
-  ghc-options: -rtsopts
+  ghc-options: -rtsopts +RTS -M512M -RTS
   build-depends: streamly-benchmarks == 0.0.0
 
 -- Some benchmarks are threaded some are not
@@ -144,7 +144,7 @@ common bench-options-threaded
   import: compile-options, optimization-options, bench-depends
   -- -threaded and -N2 is important because some GC and space leak issues
   -- trigger only with these options.
-  ghc-options: -threaded -rtsopts -with-rtsopts "-N2"
+  ghc-options: -threaded -rtsopts -with-rtsopts "-N2" +RTS -M512M -RTS
   build-depends: streamly-benchmarks == 0.0.0
 
 -------------------------------------------------------------------------------
@@ -156,6 +156,13 @@ benchmark Prelude.Serial
   type: exitcode-stdio-1.0
   hs-source-dirs: Streamly/Benchmark/Prelude
   main-is: Serial.hs
+  other-modules:
+        Serial.Generation
+      , Serial.Elimination
+      , Serial.Transformation1
+      , Serial.Transformation2
+      , Serial.Transformation3
+      , Serial.Nested
   if impl(ghcjs)
     buildable: False
   else
@@ -184,6 +191,7 @@ benchmark Prelude.ZipSerial
 benchmark Prelude.ZipAsync
   import: bench-options
   type: exitcode-stdio-1.0
+  ghc-options: +RTS -M1000M -RTS
   hs-source-dirs: Streamly/Benchmark/Prelude
   main-is: ZipAsync.hs
   if impl(ghcjs)
@@ -224,6 +232,7 @@ benchmark Prelude.WAsync
 benchmark Prelude.Parallel
   import: bench-options-threaded
   type: exitcode-stdio-1.0
+  ghc-options: +RTS -M1000M -RTS
   hs-source-dirs: Streamly/Benchmark/Prelude
   main-is: Parallel.hs
   if impl(ghcjs)
@@ -426,6 +435,7 @@ benchmark Memory.Array
 benchmark FileSystem.Handle
   import: bench-options
   type: exitcode-stdio-1.0
+  ghc-options: +RTS -M3000M -RTS
   hs-source-dirs: Streamly/Benchmark/FileSystem
   main-is: Handle.hs
   other-modules:
